Added Sign Up & reorganised files
