+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App.tsx';
+import App from './pages/App/App';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
